uart: UARTRx and UARTTx in separate files
